refactor: move JsonMovie
