+++ app/src/androidTest/java/com/aacademy/homework/data/api/FakeApiSourceImpl.kt
@@ -1,9 +1,9 @@
 package com.aacademy.homework.data.api
 
 import android.content.Context
+import com.aacademy.homework.data.api.model.JsonMovie
 import com.aacademy.homework.data.model.Actor
 import com.aacademy.homework.data.model.Genre
-import com.aacademy.homework.data.model.JsonMovie
 import dagger.hilt.android.qualifiers.ApplicationContext
 import javax.inject.Inject
