+++ app/src/main/java/com/aacademy/homework/data/api/ApiSourceImpl.kt
@@ -1,8 +1,8 @@
 package com.aacademy.homework.data.api
 
+import com.aacademy.homework.data.api.model.JsonMovie
 import com.aacademy.homework.data.model.Actor
 import com.aacademy.homework.data.model.Genre
-import com.aacademy.homework.data.model.JsonMovie
 import kotlinx.serialization.ExperimentalSerializationApi
 import javax.inject.Inject
